scout: Refactored extension for readability
