+++ Runtime/Editor/Editor UI/LootLockerAdminExtension.Authentication.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: f924e8470f005be4a8ec210a98c07efa
