+++ Runtime/Editor/Editor UI/LootLockerAdminExtension.EventHandlers.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 2bc922f0c53dab7499e00d65aa97fdec
