+++ Runtime/Editor/Editor UI/LootLockerAdminExtension.Settings.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 9654084be2f4bfb48957ad2729f182b9
